backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.115.6
filelock==3.15.4
fonttools==4.56.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pdfkit==1.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
rich==13.9.4
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.24
starlette==0.41.3
tinycss2==1.4.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.34.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
APScheduler==3.11.0
weasyprint==60.2
pydyf==0.8.0


